feature(card):  refactoring game card
